Correção das camadas
Refatoração para manter o padrão de pastas do projeto
