--- vidrelo-client/src/pages/Signin.tsx
@@ -1,13 +1,0 @@
-import Header from "../components/Signin/Header/Header";
-import Forms from "../components/Signin/Forms/FormsSignIn";
-import Footer from "../components/LandPage/Footer/Footer";
-
-export default function Signin() {
-    return (
-        <>
-            <Header />
-            <Forms />
-            <Footer />
-        </>
-    )
-}
