+++ vidrelo-client/src/styles/globals.scss
@@ -81,7 +81,3 @@ html{
     transform: translate(0);
     transition: transform 0.3cls ease-in;
 }
-
-@media screen and(max-width: $large-2) {
-    
-}
